+++ Algorithms/Create NQueen.c
