+++ Algorithms/Longest consecutive subsequence.java
