+++ Algorithms/Maximum AND value.cpp
