+++ Algorithms/Maximum difference.java
