+++ Algorithms/Painters Partition.cpp
