+++ Algorithms/Sleep Sort – The King of Laziness
