+++ Algorithms/alpha.cpp
