+++ Algorithms/binary search.cpp
