+++ Algorithms/cc.cpp
