+++ Algorithms/checkArmstrongNo.cpp
