+++ Algorithms/interpolation search.cpp
