+++ Algorithms/main (6).cpp
