+++ Algorithms/prime.cpp
