+++ Data structure/Middle Element LinkedList.java
@@ -1,0 +1,70 @@
+// Given a singly linked list of N nodes.
+// The task is to find the middle of the linked list. For example, if the linked list is
+// 1-> 2->3->4->5, then the middle node of the list is 3.
+// If there are two middle nodes(in case, when N is even), print the second middle element.
+// For example, if the linked list given is 1->2->3->4->5->6, then the middle node of the list is 4.
+
+
+
+import java.util.*;
+import java.io.*;
+
+class Node{
+    int data;
+    Node next;
+    
+    Node(int x){
+        data = x;
+        next = null;
+    }
+    
+}
+class Middle{
+	static void printList(Node node) 
+	{ 
+		while (node != null) 
+		{ 
+			System.out.print(node.data + " "); 
+			node = node.next; 
+		} 
+		System.out.println(); 
+	}
+    public static void main(String args[]) throws IOException { 
+        Scanner sc = new Scanner(System.in);
+        int t = sc.nextInt();
+        while(t > 0){
+        	int n = sc.nextInt();
+            Node head = new Node(sc.nextInt());
+            Node tail = head;
+            for(int i=0; i<n-1; i++)
+            {
+                tail.next = new Node(sc.nextInt());
+                tail = tail.next;
+            }
+            Solution g = new Solution();
+            int ans = g.getMiddle(head);
+            System.out.println(ans);
+            //printList(head); 
+            t--;
+        }
+    } 
+} 
+   
+class Solution
+{
+    int getMiddle(Node head)
+    {
+         // Your code here.
+         Node n=head;
+         int c=0;
+         while(n.next!=null){
+             c++;
+             n=n.next;
+         }
+         c=((c & 1)==0)?c/2:(c/2)+1;
+         n=head;
+         for(int i=0;i<c;i++)
+         n=n.next;
+         return(n.data);
+    }
+}
