--- Postorder Traversal.cpp
@@ -1,30 +1,0 @@
-#include <bits/stdc++.h>
-using namespace std;
-struct Node {
-    int data;
-    struct Node *left, *right;
-    Node(int data)
-    {
-        this->data = data;
-        left = right = NULL;
-    }
-};
-
-void Postorder(Node* root)
-{
-    if(root==NULL) return;
-    Postorder(root->left);
-    Postorder(root->right);
-    cout<<root->data<<" ";
-}
-
-int main()
-{
-    Node* root=new Node(1);
-    root->left=new Node(2);
-    root->right=new Node(3);
-    root->left->left=new Node(4);
-    root->left->right=new Node(5);
-    cout<<"The Postorder traversal of binary tree is \n";
-    Postorder(root);
-}
